+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 ### Changed
 
 - BC-217 - update LibreOffic domain at http-headers.js on ansible configmap-files.yml.j2
+- BC-382 - fixing error in 'removeSystem' method for new school-settings page
 
 ### Fixed